Do not render tiles when page gets AJAX loaded
