+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 3.4.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Do not render filter tiles when page gets AJAX loaded
+  [petschki]
 
 
 3.4.1 (2020-06-18)
