+++ src/collective/collectionfilter/tiles/sorting.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
-from collective.collectionfilter.tiles import DictDataWrapper
 from collective.collectionfilter.baseviews import BaseSortOnView
 from collective.collectionfilter.interfaces import ICollectionFilterResultListSort  # noqa
+from collective.collectionfilter.tiles import BaseFilterTile
 from plone.supermodel.model import Schema
-from plone.tiles.tile import PersistentTile
 from zope.interface import implementer
-
-import plone.api
 
 
 class ISortOnTile(Schema, ICollectionFilterResultListSort):
@@ -14,25 +11,5 @@ class ISortOnTile(Schema, ICollectionFilterResultListSort):
 
 
 @implementer(ISortOnTile)
-class SortOnTile(PersistentTile, BaseSortOnView):
-
-    @property
-    def edit_url(self):
-        if not plone.api.user.has_permission(
-            'cmf.ModifyPortalContent',
-            obj=self.context
-        ):
-            return None
-        return self.url.replace('@@', '@@edit-tile/')
-
-    @property
-    def settings(self):
-        return DictDataWrapper(self.data)
-
-    @property
-    def filter_id(self):
-        return self.id
-
-    @property
-    def reload_url(self):
-        return self.url
+class SortOnTile(BaseFilterTile, BaseSortOnView):
+    """ Sorting Tile """
